# gest-ashoex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Health Check",
        "description": "API para verificar el estado de salud del sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ubicaciones": {
            "get": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Obtiene todas las ubicaciones con sus aulas",
                "operationId": "740fdf71103b541cb086cae7b576d096",
                "responses": {
                    "200": {
                        "description": "Lista de ubicaciones recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_ubicacion": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "piso": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "id_edificio": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "aulas": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id_aula": {
                                                                    "type": "integer",
                                                                    "example": 11
                                                                },
                                                                "numero_aula": {
                                                                    "type": "string",
                                                                    "example": "INF123"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de ubicaciones recuperada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ubicaciones"
                ],
                "summary": "Almacena una nueva ubicación en la base de datos",
                "operationId": "95941fd74578a777e4b31c4bc0e4cc26",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "piso",
                                    "id_edificio"
                                ],
                                "properties": {
                                    "piso": {
                                        "description": "Número de piso de la ubicación",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "id_edificio": {
                                        "description": "ID del edificio donde se encuentra la ubicación",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ubicación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "piso": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "id_edificio": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_ubicacion": {
                                                    "type": "integer",
                                                    "example": 9
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ubicación registrada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "422"
                                                    },
                                                    "detail": {
                                                        "type": "string",
                                                        "example": "El campo piso es obligatorio."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "status": 422,
                                                    "detail": "El campo piso es obligatorio."
                                                },
                                                {
                                                    "status": 422,
                                                    "detail": "El campo id edificio es obligatorio."
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}